Disable Shift as replacement for Alt (Option) on Linux and Mac OS X
Shift was used as replacement for Alt to allow moving backward (issue #40), but
it turned out not to be a good solution.

Now to move backward use:
- Option + Left button mouse (LBM) on Mac OS X
- Alt + LBM or Alt + Meta (WinKey) + LBM on Linux, depending on your desktop environment

This fixes issue #123 and reverts changes made for issue #40.

If there are still users of one-button Mac mouse, you are out of luck, sorry.
You should start the game with flag -no3btnmouse, that will allow you to use
Command + mouse click as replacement for right mouse button, but you won't be able
to move backward.