Fix #198: end of meanwhile cutscene crash
If a cutscene is ended by an NPC action then an extraneous dialogue pop
is executed. This patch adds a check to prevent an empty queue from
being acted on. The true bug may be the return value of Execute and the
correct patch to return the opposite boolean value.