Fix mouse moving out of the windows when in fullscreen mode
When the engine is configured to be in fullscreen and the size of
the window is smaller than the desktop resolution, the mouse moves out
of the window when it goes over the edge. And, if your desktop is like
mine configured to have focus follow mouse, the game windows immediately
loses focus, which makes the game pretty much unplayabl.

Now, to avoid this, it's possible to set the size of the window at the
same resolution as the desktop but, with the size of today's screens, it
makes the objects/characters pretty small.

So, I fixed this by adding the flag SDL_WINDOW_INPUT_GRABBED to the
window when it's in fullscreen so the cursor doesn't wander out of the
window. And, of course, when switching back and forth to
windowed/fullscreen mode, I needed to add a call to SDL_SetWindowGrab to
set/unset that flag: if the flag is set when in windowed mode, the
cursor can't leave the window, which I think isn't a desirable behavior.