Specify output of Rust-dependency
The ja2 binary depends on libstracciatella.a, which the Ninja build
system expects to either be available or have a recipe for.
The latter doesn't exist and the former doesn't hold when the build
starts.
We tell the externalproject that it produces the library so CMake
generates a build-target for it and Ninja is happy.