Set USES_TERMINAL for Rust build
This puts the task in Ninja's console-pool, which produces immediate
output instead of buffering.
The Rust build does take a little while and this will report the
progress to the developer instead of silently blocking.

This also has the consequence of stunting parallel task execution,
though since Cargo itself provides parallel builds this is not an
issue.