Untangle the dependencies of module config.
The config submodules depended on symbols reexported in the root.
The root symbols depended on symbols reexported in the config module.
The config module reexported symbols of the non-public submodules.

Since the config submodules are not public, it's fine to reexport their
symbols in module config to allow public access.