Convert canonicalized UNC paths to normal paths when possible
On windows the launcher couldn't find the path to game.json because the
canonicalized base path was in UNC form and the relative part had a /
path separator.

With a normal base path it works as expected, so crate dunce is being
used to turn UNC paths into normal paths when possible.