Make GetAPsToAutoReload agree with AutoReload
Coverity detected "Negative array index read (NEGATIVE_RETURNS)" in
GetAPsToAutoReload when we can't find ammo for the first gun and check
if that ammo is compatible with the second gun.

AutoReload only reloads the second gun if the first one reloads so here
we enforce the same behavior, avoiding the check.