Build rust code only when needed
Trigger a rebuild of stracciatela when the sources change.

The md5 checksums and paths of the source files are computed every time
cmake runs.

When the computed data is different from the data in the md5 file, the
md5 file is updated and the update file is touched if it exists.

The update file is the stamp for the update step of the external
project.