Fix out of bounds reads/writes
Coverity detected at least one out of bounds read in ManSeesMan when the opponent id is 155.

SOLDIERTYPE::bOppList has length 148 and is part of the AI system.
Ids 148 to 155 belonged to the PLAYER_PLAN team, used in the planning mode.
The planning mode was removed by tron because it was unused (not bound to a hotkey).
These soldiers were temporary and only existed during the player turn, they were not meant to have AI interaction.

This commit fixes the range checks that are related to arrays with length 148, there might be more.