Reimplement cinematics.
It uses libsmacker directly.

It supports playing 4 videos at the same time.
Before it had 4 slots but could only play correctly 1 at a time.

It supports playing all sound channels.
Before it only played the first channel.

It does not desynchronize over time.
It synchronizes using an initial tick.
The start time of the sound and the video is still a problem.