Fix unintended sign extension in BlitPcxToBuffer.
Coverity detected an unintended sign extension in BlitPcxToBuffer.

In the w * h operation, w and h are first promoted to int, then multiplied, then sign-extended to size_t.
When w * h is greater than 0x7FFFFFFF, the int will be negative.
When size_t is 64 bits and the int is negative, the upper bits will be 1. (bigger than UINT32_MAX)

UINT16_MAX * UINT16_MAX = 0xFFFE0001, so a bigger value is a wrong result.