Replace memset in SoundMan.cc and fix issues
Due to a mixup between elements and bytes, the mix buffer was created with 4 times the needed size.
Now it allocates the exact amount.

GetSampleSize returned size_t, which caused uneeded conversion warnings.
Now it returns UINT32.

A negative len would lead to underflows.
Now it returns immediately.

A len that wasn't a multiple of 2 would leave 1 uninitialized byte.
Now it initializes everything, as the docs require.