Allow known map versions
Non-russian maps can be loaded when playing the russian version of the game.
Russian maps are loaded correctly when playing a non-russian version of the game.
Newer (unknown) versions are not loaded.

The map version is checked against the minor version.
A missmatch implies a bad load.

Warn when an unexpected version is loaded.

Include the description of the exception when setting the error.