Fix unintended fallthrough in HandleSummaryInput
After testing, it's probably an unintended fallthrough.

When you start the editor it mentions quitting with ESC or ALT+x.
Before loading a map, ESC quits the game (expected).
After loading a map, ESC had the same effect as ENTER or clicking the Okay button.
Now it ignores ESC after a map is loaded.