Remove extra warnings
I tried to fix "enumeral and non-enumeral type in conditional expression", but I coun't find of think of a rationale for the warning.
Are they trying to say "stop using C-style enums"?
That's unlikely to happen until there is an auto-numering alternative.

There are three ways to handle the warning:
1) cast enum value to an integer type
2) convert enum to constants
3) remove option -Wextra

The -Wextra warnings seem to be subjective or a matter of taste.
It's a waste of time to worry about them when we don't even have a preferred style.
Anything that actually matters should be covered by coverity, so I'm removing the options related to -Wextra.