Skipping placing invalid items placed on maps
If the item has a ITEM_NOT_EDITOR flag, it is not supposed to be placed on the map.

Some maps (e.g. Wildfire maps) are saved for different item set. We should ignore and print a warning, rather than placing a weird unusable item.

The "creature gas" items are redefined in Wildfire, but on Vanilla, they are internal items that should not be placed.