Ensure Quick Edit mode is off (Windows only)
The Windows console feature "quick edit mode" freezes the STDIO when users clicks on the console - so the user can select and copy the text - but that also pauses program execution.

This seems to default to on in Windows 10, and we should disable it on startup as it likely causes confusion if user accidentally clicks on the console.

User can still manually enable the quick edit mode from the menu.