Do not display hatch on items being repaired
The check we had before always evaluated to FALSE, so we never had a hatch for items to be repaired.

The hatch display would make the repair ETA unreadable. Completed items are moved to the bottom panel of the SKI, and no longer require a hatch. In all cases the hatch should not be displayed for repairs.