AppVeyor builds to use VS 2019
There is an issue with std::current_exception that prevents us from logging unhandled exceptions. The bug is fixed in Visual C++ 2019. See MSConnect 3127473 (https://developercommunity.visualstudio.com/content/problem/135332/stdcurrent-exception-returns-null-in-a-stdterminat.html)

The CMake generator of VS 2019 no longer specifies the platform, and the platform must be provided separately via '-A' or env.