Use fixed Windows SDK version on appveyor
According to https://developercommunity.visualstudio.com/content/problem/1139955/c-builds-with-windows-sdk-81-fail-after-upgrade-to.html there is an issue with windows sdk in version 8.1 and Visual Studio 16. Since those are the versions currently in use on appveyor, my hope is that this does not happen with windows sdk 10.0, as written in the linked issue.