sgp: added SGPSector class and started converting users
The idea is to reduce the number of parameters and globals, simplify
comparisons and setting, streamline special behaviour, and reduce LOC.
Currently we are are still doing several temporary object creations, so
benefits in brevity will only start coming online as the conversion
comes closer along.

Some functions don't care about the Z coordinate, but the defaults
are set so that Z doesn't get in their way. Users assume all sorts
of signedness for the coordinates, but so far all deviations have
been inconsequential.