Remove SGPVSurfaceAuto
The only difference between SGPVSurfaceAuto and SGPVSurface is that the
former frees the underlying SDL_Surface upon destruction while the latter
does not. There is only one surface where we do not want this automatic
free to happen: the screen buffer.

This behaviour was added back in 2013 to fix issue #45: crashes on MacOS.
So why do we not need this special behaviour any more? Back then the
wrapped surface for the screen buffer was obtained by a call to the
SDL 1.2 function SDL_SetVideoMode, and that surface must not be freed
by the application, similar how the surface obtained by a call to
SDL_GetWindowSurface in SDL 2.0 must not be freed.

Nowadays Stracciatella uses CreateRGBSurface for the screen buffer, so
the need for SGPVSurfaceAuto simply is not there any more. SGPVSurface
can always free its surface.