Do not update the entire screen texture every frame
Before this commit, RefreshScreen() used SDL_UpdateTexture to blit the
entire screen buffer surface to the screen texture every frame. Since
this function has to convert the screen texture format to the actual output
format of our window, this is extremely costly. The game spent well above
50% of its time under this function.

With this commit, only the union of all screen buffer regions that actually
got modified are updated. In the most extreme cases (Load Screen and others)
this union can be smaller than 1% of the screen area.