Fix assertion failure in GetSelectedAssignSoldier
DetermineWhichAssignmentMenusCanBeShown must not call this function
before checking some flags, otherwise mousing over a vehicle triggers
a stream of assertions.

This commit includes another fix for a problem found by clang-tidy:
being a POW was supposed to be stressful and impede healing, but this
never worked due to a missing 'else'.