Change how often GameLoop() gets called
Before this commit, the duration between two GameLoop calls was
configurable (ms_per_game_cycle, default 25ms). This is problematic
because many of the predefined timecout counters use multiples of 5ms,
and the delay between two soldier animations is even calculated with
a resolution of 1ms.

With this commit the game now tries to call GameLoop at a rate of
144Hz (every ~7ms), corresponding to the refresh rate of modern
monitors. To satisfy the needs of our users who do not need 144 FPS
and would prefer a lower CPU usage, GameLoop()'s heaviest subfunction
RefreshScreen() now has its own timeout. This timeout will be made
user configurable with the next commit.