Fix division by 0 in CalcBestStab()
MinAPsToAttack() can return 1 for extreme values of ubShotsPer4Turns
in weapons.json, so always subtracting 1 AP for the cost of turning
even if the soldier does not need to turn for the attack gave 0 for
ubRawAPCost and a division by 0 followed when computing the hit rate.

CalcTentacleAttack() had the same problem.

Removed declaration of non-existant function GetBPsTouseJar().