Give special treatment to some keys used as shortcuts in tactical
Treat the keys that are labeled with the grave, equals sign and slash
on an US keyboard like those keys on all keyboard layouts so they
can be used as shortcuts in tactical mode. Does not affect text input
when entering save game descriptions.

Also add ',' as an alternative to '*' to toggle red item glow.

Fixes #1844.