Fix assigning mercs to Repair in tactical
RepairMenuBtnCallback got the merc who should do repairs from
a global variable, which could get clobbered between the time
the popup box was created and the time the callback function
was executed.

Now the ID of the merc is stored in the user data of the popup
boxes' mouse regions during creation and retrieved from there
by the callback function. Issue #1850.