Improve loading strings from EDT files
This commit adds the following:
1. A new interface IEDT. This very simple interface has
just one method to get a single string from an EDT file.
2. Class ClassicEDT is a helper class for files in the
original .edt format. It does not do much, it simply
opens the file and takes care of offset calculations.
3. Class JsonEDT is only available for mods via the
ModPackContentManager. It uses a classic .edt file as a
fallback option and allows to override individual strings
with entries taken from a .json file.
4. Class EDTFile is the intended replacement for the
current loadEncryptedString calls. It contains the details
needed to interpret the layout of the supported .edt files.