Remove the limitation that all weapons must have an index <= 70
The only code reason for this limitation was the size of the
STRATEGIC_STATUS::fAlreadyDropped array which used one BOOLEAN
per weapon.

This commit utilizes the formerly unused passing bytes at the
end of STRATEGIC_STATUS and uses one bit per weapon for a new
theoretical limit for the weapon's index of 629. The total
number of items is still limited to MAXITEMS (351), however.