Bug fix: size of SaveStates in the save game header
The size of ste SaveStates was written to the file header
at the start of SaveGame but since #2051 was checked in,
saving the strategic status added more data to those
SaveStates. The result of this bug was that SaveGameInfo
could no longer determine the mods activated for a save,
resulting in wrong tool tips on the load screen and wrong
warnings about a mismatch in the activated mods.