Move the FPS limit out of RefreshScreen to a separate function
RefreshScreen was originally intended to force a refresh of
all regions that were marked as invalid at the time. On
systems more modern that the original minimum system
requirements that resulted in hundreds of FPS and high
CPU usage so it got a limiter. However, this meant that in
some situations the screen was not updated often enough,
resulting in somewhat choppy animations or content that
might not show up at all.