Remove workaround for an old SDL bug in ScrollJA2Background
Older versions of SDL_BlitSurface did not always handle the
case where the source and destination surface are the same
correctly. This bug has been fixed quite a while ago and is
included in SDL 2.0.5, which is coincidentally also our
current minimum requirement so we can remove the workaround
where we use a temporary surface and blit twice.