Fix warning C4146 in SaveGameInfo
MSVC correctly points out that the combination of unary
minus and an unsigned integer does not do what one might
expect; the result is still positive. In our case we are
passing a huge number close to 2^32 to seek() and the only
reason we are getting the expected result is that seek()
interprets this number as a small negative number. If we
would ever modernize seek's first argument to int64 this
would break.