Revive the radar map utility
This commit brings back the utility screen to create radar
maps for all map files found in the user's home directory
that was removed in #1300.

Also reverts part of #1706.

I chose to make this utility available the same way the
1.13 editor does it because I assumed modders are likely
to be familiar with it: by adding a new button on the
editor's options tab. After clicking that button the
utility does its job and then exits the program.