Fix crash in StartInterrupt()
In some rare situations, this function can get called when
none of the members of one of involved teams can really act
(either dying or out of breath) which means that its AI
list would be empty. StartInterrupt() didn't handle this
and just tried to use the first entry of the AI list and
then crashed when it tried to deref a NULL pSoldier.