Fix surprising conversion of string literals to JsonValue
MercProfile::serializeStruct contains several lines like
this:

obj.set("074sexismMode", "GENTLEMAN");

where obj is a JsonObject and set's second parameter is a
JsonValue. One would expect this line to set the property
"074sexismMode" to the string "GENTLEMAN" but with the old
set of JsonValue constructors this would actually choose
the constructor taking a bool value, so in effect this line
set the property to a bool with the value true.

Found by clang and Coverity (with a quite confusing
diagnostic message), CID 419742.