Properly center the history table header columns (#2136)
* Properly center the history table header columns

The three column headers were all supposed to be centered
but the 'Event' header was not since the the vanilla days.

The imminent cause of this bug is almost certainly a case
of copy & paste and then forgetting to adjust the arguments
but I believe the real underlying problem is the text API
which makes this often needed task far more cumbersome to
use than necessary:

The first method is to use FindFontCenterCoordinates or
FindFontRightCoordinates followed by MPrint. This method
requires passing several arguments twice.

The second method is the DrawToScreen function. This is
just one call but it doesn't respect the currenly active
font and colors and instead requires the caller to always
pass in this information.

For this reason this commit adds a new MPrint overload that
takes an additional argument to specify the alignment.