Use SDL to simplify AddVideoSurfaceFromFile
In its original form CopyImageToBuffer supported copying
compressed data to a surface. It lost that ability long ago
apparently because it was not used and now this function is
just an elaborate way to copy raw data from a SGPImage to a
SGPVSurface.

SDL_ConvertPixels can do this copy just as well which allows
us to remove CopyImageToBuffer and its helper functions.