Properly reset gfPendingButtonDeletion
There is a mechanism to ensure that button callback handlers
can safely can RemoveButton. If this happens, the button is
not deleted immediately but only marked for deletion and the
flag gfPendingButtonDeletion is set.

At the end of button system's mouse region callback handler,
this flag is checked and if it is set, the function
RemoveButtonsMarkedForDeletion is called which finally
deletes all marked buttons.

However the flag was never reset which means the function
was called every time even when it is no longer necessary.