Replace FOR_EACHX with ranged for loops
The FOR_EACH and FOR_EACHX macros predate C+11 and were
intended to iterate over C style arrays. FOR_EACHX tries
to mimic the normal for statement by allowing an additional
expression to execute at the end of the loop. This looks
a bit strange because all parts of the for loop are
seperated by commas.

Using an ordinary ranged for loop and moving the extra
expression inside the loop body improves readability.